+++ .eslintrc.js
@@ -1,3 +1,15 @@
 module.exports = {
-    extends: ['@polarislabs/eslint-config-polaris/node']
-}
+    "env": {
+        "commonjs": true,
+        "es6": true,
+        "node": true
+    },
+    "extends": "eslint:recommended",
+    "globals": {
+    },
+    "parserOptions": {
+        "ecmaVersion": 2018
+    },
+    "rules": {
+    }
+};